Android 6.0.1 update
Fix for customwebview render. In newest android target SDK+ (android v13 or higher), the webview requires that you manually change the properties "AllowFileAccess", "AllowFileAccessFromFileURLs" and "AllowUniversalAccessFromFileURLs" if you're pointing a webview to a resource on disk (even if you explicitly have access to that location), the webview will currently DEFAULT to an error unless you make these changes.